+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Slightly corrected Google Drive links
+* Use `pathlib` to work with paths
 
 ## Release 1.1 (2021-04-09)
 * Added an icon (it is Base64 encoded in the code, instead of downloading it)